requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
aiosqlite==0.19.0
SQLAlchemy==2.0.29
python-dotenv==1.0.0
flask==3.0.0
flask-session==0.6.0
flask-cors==4.0.0
requests==2.31.0